GCS_MAVLink: add support for mission planning in guided and auto mode
Since retirement of the AVR version, MAVProxy and other GCS have adopted
the 2.0 version of MAVLink protocol. As it is currently not possible to
bring all new features from modern Ardupilot into the narrow 1MB flash
of the Atmega2560, we can, however, adapt by ignoring the extra fields
from mav20, making it possible to receive messages that are
understandable and execute supported commands.

In the particular case of the deprecated MSG_ID_MISSION_ITEM in favor of
MSG_ID_MISSION_ITEM_INT we ignore `mission_type` and convert into a
mavlink_mission_item_t packet that is readable by the older atm version.

For that purpose, it was necessary to generate new interfaces using the
`mavgen` script, while narrowing it down to use minimum memory as
possible (6456711f).

Signed-off-by: Saulo G. Felix <meetnapoleon@protonmail.ch>