AP_Mission: fix mission planning on the APM 2.X
Modern GCS send and expects to receive a mission item of type
mission_item_int_t. A lot of the underlying mechanismis relys on
structs of type mission_item_t which treats lat/lng and alt values
as float.

A simple way to make it compatible while avoiding introducing gps
glitches and other bugs is to momentarily parse mission requests
and mission items of type int as float then following command execution
as usual, until we have to acknowledge or send values back as mavlink
messages, in these cases we convert it back to mission_item_int_t as
expected by the GCS.

Signed-off-by: Saulo G. Felix <meetnapoleon@protonmail.ch>